Move Blob descriptor resolver out of writer
